Added RecorConnectionIdentifier to Measurement Changed value_as_number from an int to a double

+++ OmopTransformer/Omop/Measurement/OmopMeasurement.cs
@@ -9,11 +9,11 @@ internal abstract class OmopMeasurement<T> : IOmopRecord<T>
     public virtual DateTime? measurement_time { get; set; }
     public virtual int? measurement_type_concept_id { get; set; }
     public virtual int? operator_concept_id { get; set; }
-    public virtual int? value_as_number { get; set; }
+    public virtual double? value_as_number { get; set; }
     public virtual int? value_as_concept_id { get; set; }
     public virtual int? unit_concept_id { get; set; }
-    public virtual float? range_low { get; set; }
-    public virtual float? range_high { get; set; }
+    public virtual double? range_low { get; set; }
+    public virtual double? range_high { get; set; }
     public virtual int? provider_id { get; set; }
     public virtual string? measurement_source_value { get; set; }
     public virtual int? measurement_source_concept_id { get; set; }
